+++ backend/canisters/group/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add support for sending the KINIC token ([#3811](https://github.com/open-chat-labs/open-chat/pull/3811))
 - Support replying to channel messages ([#3825](https://github.com/open-chat-labs/open-chat/pull/3825))
 - Support simplified transfers of icrc1 tokens ([#3827](https://github.com/open-chat-labs/open-chat/pull/3827))
+- Implement converting a group into a community ([#3833](https://github.com/open-chat-labs/open-chat/pull/3833))
 
 ### Changed